requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
cryptography>=43.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
watchfiles>=0.24.0
supabase>=2.10.0
httpx>=0.27.0
pydantic>=2.0.0
gunicorn>=22.0.0
pytest>=8.0.0
